fix migration hook
